LayoutTests/imported/w3c/web-platform-tests/wasm/core/instance.wast.js-expected.txt


PASS #1 Reinitialize the default imports
PASS #2 Test that WebAssembly compilation succeeds (instance.wast:3)
FAIL #3 Test that WebAssembly compilation succeeds (instance.wast:15) assert_equals: expected false but got true
FAIL #4 Test that WebAssembly compilation succeeds (instance.wast:62) assert_equals: expected false but got true
PASS #5 Test that WebAssembly compilation succeeds (instance.wast:109)
FAIL #6 Test that WebAssembly compilation succeeds (instance.wast:128) assert_equals: expected false but got true
PASS #7 Reinitialize the default imports
PASS #8 Test that WebAssembly compilation succeeds (instance.wast:3)
PASS #9 Test that WebAssembly instantiation succeeds (instance.wast:3)
PASS #10 Test that WebAssembly instantiation succeeds (instance.wast:3)
FAIL #11 Test that WebAssembly compilation succeeds (instance.wast:15) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 95: there can at most be one Memory section for now at {loc} expected true got false
FAIL #12 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:25:18
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #13 Test that a WebAssembly code returns a specific result (instance.wast:54) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #14 Test that a WebAssembly code returns a specific result (instance.wast:55) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #15 Test that a WebAssembly code returns a specific result (instance.wast:56) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:34:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #16 Test that a WebAssembly code returns a specific result (instance.wast:57) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:37:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #17 Test that WebAssembly compilation succeeds (instance.wast:62) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 95: there can at most be one Memory section for now at {loc} expected true got false
FAIL #18 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:43:18
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #19 Test that a WebAssembly code returns a specific result (instance.wast:101) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:46:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #20 Test that a WebAssembly code returns a specific result (instance.wast:102) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:49:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #21 Test that a WebAssembly code returns a specific result (instance.wast:103) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:52:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #22 Test that a WebAssembly code returns a specific result (instance.wast:104) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:55:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
PASS #23 Test that WebAssembly compilation succeeds (instance.wast:109)
PASS #24 Test that WebAssembly instantiation succeeds (instance.wast:109)
FAIL #25 Test that WebAssembly compilation succeeds (instance.wast:128) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 95: there can at most be one Memory section for now at {loc} expected true got false
FAIL #26 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:72:18
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #27 Test that a WebAssembly code returns a specific result (instance.wast:167) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:75:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #28 Test that a WebAssembly code returns a specific result (instance.wast:168) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:78:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #29 Test that a WebAssembly code returns a specific result (instance.wast:169) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:81:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
FAIL #30 Test that a WebAssembly code returns a specific result (instance.wast:170) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
instance_wast_js@http://web-platform.test:8800/wasm/core/js/instance.wast.js:84:14
global code@http://web-platform.test:8800/wasm/core/js/instance.wast.js:86:3 expected true got false
